Unit conversion of Interface values (InterfaceType must define a Unit)
On InterfaceCommand execution, values are converted to the reference unit of the InterfaceType. If the value is an expression, the conversion is stated symbolically: "({Value}) * {str(conv_factor)}".

To ensure InterfaceTypes have a unit, the field Unit of InterfaceType command is defined now as mandatory.

Also, bug solved regarding case sensitiveness of DatasetDef, DatasetData commands
Also, reenumeration of some tests.